Implemented basic jump instructions and if statements
Currently the instructions use a 5-byte operand allowing a jump of size 2^40 - this may be overkill since that's about a terabyte of bytecode but equally I felt the 16MB addressable by a 3-byte operand could be too restrictive.

I may revert to a 3 or 4 byte operand if this seems to affect performance particularly, though I can't really test that until loops have been implemented.