Implemented break and continue statements
These are tracked through a new loop stack that is attached to the compiler. This stores the chain of loops that have lead to the current location, tells continue statements where to loop to (with the exception of do-while) and keeps a record of breaks that need to be patched at the end of the loop.

The loop stack can also feature sentinel values which will be more useful when functions are implemented (the idea being to ensure a function being compiled inside a loop can't break/continue the loop).