Native functions can now raise runtime errors
This is done through the addition of extra type to signify native errors but one that should never appear on the stack.

This allows native functions to perform their own type and arity checking and raise errors should anything be incorrect. I leave arity checking to the callee since this allows native variadic functions (e.g. piping through to printf) "for free"