Fixed memory issues with import system
Forgetting a break was embarrassing but the main issue was with string interning. Because secondary VMs had weak references to their parents' string table, this meant they'd free all the strings not in their scope whenever gc fired, breaking *everything*. The least horrible fix I could come up with was just to disable string gc during an import, which should only be an issue if someone writes a library that does a lot of direct string manipulation, which is unlikely.