Removed old native functions for type conversion as well as println
The old conversion functions could now not be reached and were entirely inferior to those that replaced them.

println had to be removed since it's main advantage (implicit conversion of arguments to string) could no longer be realised as conversion to string is not possible in a native function. This is because it is impossible to patch out to an override function written in canidae from native C code.

To compensate for this, a print statement implicitly converts the final expression to the right of it to a string using OP_CONV_TYPE.

Tests also had to be changed to deal with the removal of println